+++ activerecord/lib/active_record/connection_adapters/sqlite3/database_statements.rb
@@ -138,7 +138,11 @@ module ActiveRecord
           end
 
           def default_insert_value(column)
-            column.default
+            if column.default_function
+              Arel.sql(column.default_function)
+            else
+              column.default
+            end
           end
       end
     end