+++ activerecord/CHANGELOG.md
@@ -1,3 +1,5 @@
+## Rails 8.0.1 (December 13, 2024) ##
+
 *   Fix removing foreign keys with :restrict action for MySQ
 
     *fatkodima*